Re-introduce mutation groups
Re-introduce mutation groups without the usage of interfaces,
only using built-in intefaces.

The reasoning for re-introducing the grouping of mutations is that
it is possible to exclude mutations in groups, in case a single
mutation in the group caused a time out.

Whether this is desireable design is beyond the scope of this work.